fix: add convertion to int for new fields

+++ src/modules/schedule/utils/index.ts
@@ -76,8 +76,8 @@ export const buildScheduleQuery = (whereClause: SQL[]): SQL<unknown> => {
       row_number() over (
         partition by s.id, e.type
         order by e.started_at
-      ) as "pairIndex",
-      count(*) over (partition by s.id, e.type) as "pairsCount"
+      )::int as "pairIndex",
+      count(*) over (partition by s.id, e.type)::int as "pairsCount"
     from
       event e
     join auditorium a on a.id = e.auditorium_id
